Create README.md
Originally, the robot was supposed to be a tree whose personality would exist in his branches moving. He was intended to have less emotion and not have as much of a personality. However, at the final performance, the mirrors attached to servos ended up looking like two eyes. So at the final performance, I noticed that he had been anthropomorphised a bit. The lights going out as the performance progressed helped show that he was going to sleep, falling victim to the nihilistic thoughts that he was ‘voicing’. The personality was supposed to come mainly from the voice, however, the design and effects ended up humanizing ASAP Tree accidently. While having a cool DJ who was more mechanical and distant(like Daft Punk or DeadMaus), ASAP Tree ended up having a cute, quirky face that twirled around and slowed down as time went on because he was being emotionally affected by the philosophical struggles he was voicing.
The original idea was to create a tree. It would have had three rotating chassis made out of three different materials(wood, acrylic, and aluminum). These would rest on a CNC milled wooden table. However, getting the chassis to rotate turned out to be a nightmare. The aluminum in particular was incredibly difficult to screw together as well as rotate. This could have potentially been solved by making the rings thicker, but the circular nature of the chassis made it difficult to maintain their shape while rotating and to follow the track. We should have scrapped the idea of the rotating chassis much earlier than we did, and we wasted a lot of time trying to force things. The end product of using wood to lift the rotating laser box ended up being much more stable and workable. If I could go back, I would make a more unique body shape, similar to a tower to build a body for the DJ. I would utilize the mirrors looking like eyes and try to make something more human like, and paste the neopixels into more interesting shapes than just being on the wooden platform. I would never have touched the aluminum in hindsight either. It was difficult to saw and drill into and just all around not viable at all. I would have also been more flexible in my design. I got stuck in a rut of thinking that the robot had to follow our original plan, and when Arb and I didn’t have the skills to pull that off, we didn’t accept that and tried to cobble together what we could. We should have scrapped the base and worked on the pixels and lasers from the beginning and strengthened those more rather than work on the chassis. 
I had to learn how to use power tools and Adobe Illustrator to make the chassis. I learned how to make CAD models in Tinkercad as well. All of this was through James’s help. I also had to learn about neopixels, as I had never used them before either.  I got a lot of help from the built in libraries and just reading over the examples on stack overflow in order to get a good grip on the general logic and commands used in the neopixel library. I think that one of the more difficult parts of the project was learning to use materials that weren’t cardboard and glue, because that’s what I had exclusively been using up until this point.

The project has three main electrical components: the neopixels, the rotating box, and the servo mirrors. Each object is set up with the bare minimum amount of wiring. The neopixels were both connected to Arduinos with motor shields and bluetooth connectors. The rotating box was just a battery pack, switch, and motor. The servos were controlled with an Arduino to regulate their speeds and had a switch to connect the battery packs and turn on the servos. Each of these individual pieces was kept separate and had no combined power sources or were connected in any way. 

The software of our project mainly consisted of the bluetooth controlled neopixels. There were four buttons that escalated in the colors, brightness, and frequency as you went higher. The neopixels had several different effects from a few dots going across the rings too waves of lights illuminating all at once. The servos holding the mirrors were also set to go at 55 by an Arduino so that the mirrors wouldn’t fall off.

The mechanical aspects of the project mainly consisted of the rotating laser box. We used the laser cutter to create the box, and inside glued down the servos on lifts that would hold the mirrors at the correct angles. We also had to drill through the acrylic to mount the lasers. The acrylic ended up cracking and we had to use glue, tape, and zip ties to compensate. The lasers were held down by tightening clamps to fit loosely over the laser’s buttons and then placing cardboard inside to press the button down. The box rested on a large wheel being turned by a motor clamped to an L-shaped piece of wood that elevated the mechanism. The neopixels were soldered together and glued/taped to the side of the larger base. We used a CNC miller to make the large base, using a design made in Adobe Illustrator. There was also a lot of mechanics involved in measuring and designing the chassis, not to mention sawing and drilling them all together, however, they weren’t used in the final project so that ended up being a wash.

The three most difficult parts of the project were letting go of the chassis idea, working as a team, and putting the neopixels together last minute. We should have restructured our idea much earlier on, and we didn’t. I just couldn’t seem to say to myself that this wasn’t working and find an alternative. Lesson learned though, I think I would have been much more content with a different project instead of wasting so much time on trying to force an idea that wasn’t playing out. Working as a team was also a challenge. Communication between me and my partners wasn’t where it should have been, and we wasted a lot of time and energy re-doing each other’s work because we either didn’t communicate that a job had been completed or weren’t clear with our instructions on what we were envisioning. In the future, a plan should be much more well threshed out in group projects of this nature. Lastly, putting the neopixels back together last minute added a lot of stress, and I think that made me more critical of the final performance. I would have enjoyed the Robot Liberation Theater more if I hadn’t been as disappointed that I didn’t make the robot that I had envisioned. I also should have been more careful about checking voltage levels before plugging in the neopixels. The math involved in building those circuits should have been given more careful consideration, and I skated over going more in depth to that aspect of the class.

In hindsight, I would have been a better planner, changed my design after encountering flaws and hinderances that I couldn’t have fixed, and had a better minimum viable option. I think I tried to do too much with too little experience and I was too stubborn in waiting to admit that I couldn’t produce the product I wanted too. I think that I need to be more flexible in changing my plans in the future and looking for easier alternatives instead of getting stuck in the same rut.